Add `TraceEventLogger` to the compile time layer.
- Move `TraceEventLogger` and its helper functions to `layer/support`.
- Implement `LayerDataWithTraceEventLogger`, a wrapper that adds the new
  logger to `LayerData`.
- Add `TraceEventAttr` to all of the compile time layer's events.
