+++ layer/support/common_logging.h
@@ -29,10 +29,8 @@ namespace performancelayers {
 // `event_name,attribute1_name:attribute1_value,attribute2_name:attribute2_value,...`
 std::string EventToCommonLogStr(Event &event);
 
-// CommonLogger logs the events in the common log. There is no need to add '\n'
-// at the end of the input of the `AddEvent()` method. This is handled by the
-// implementation. The only valid methods after calling `EndLog()` is
-// `EndLog()`.
+// CommonLogger logs the events in the common log. The only valid method after
+// calling `EndLog()` is `EndLog()`.
 class CommonLogger : public EventLogger {
  public:
   CommonLogger(LogOutput *out) : out_(out){};
